% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_krasddpcams.R
\name{create_krasddpcams}
\alias{create_krasddpcams}
\title{Main function}
\usage{
create_krasddpcams(base_dir = "./")
}
\arguments{
\item{base_dir}{Base directory for all output files (default:current working directory)}
}
\value{
Nothing
}
\description{
This is the top-level function that reproduces all figures and output files.
}
