% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doubledeepms__abundance_predict_fitness.R
\name{doubledeepms__abundance_predict_fitness}
\alias{doubledeepms__abundance_predict_fitness}
\title{prediction abundance fitness from folding energy change}
\usage{
doubledeepms__abundance_predict_fitness(
  mochi_parameters = mochi_parameters,
  fold_n = fold_n,
  folding_energy = folding_energy,
  RT = 0.001987 * (273 + 30)
)
}
\arguments{
\item{RT}{R*T}

\item{input}{dimsum data}

\item{assay_name}{folding or binding}

\item{anno_input}{anno_input}
}
\value{
predicted abundance fitness
}
\description{
This function allows you to plot fitness distribution.
}
