% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/doubledeepms__minimum_interchain_distances_from_PDB.R
\name{doubledeepms__minimum_interchain_distances_from_PDB}
\alias{doubledeepms__minimum_interchain_distances_from_PDB}
\title{minimum interchain distance Function}
\usage{
doubledeepms__minimum_interchain_distances_from_PDB(
  input_file,
  chain_query = "A",
  chain_target = "B"
)
}
\arguments{
\item{input_file}{PDB file}

\item{chain_query}{KRAS chain}

\item{chain_target}{binding interactor chain}
}
\value{
result_dt
}
\description{
This function allows you to calculate minimum interchain distance from PDB files.
}
