% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Get_binidingresidue_bindingpartners.R
\name{Get_bindingresidue_bindingpartners}
\alias{Get_bindingresidue_bindingpartners}
\title{Get binding residues of binding partners}
\usage{
Get_bindingresidue_bindingpartners()
}
\value{
Nothing
}
\description{
This function allows you to get binding residues of binding partners.
}
