% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Get_combined_threshold.R
\name{Get_combined_threshold}
\alias{Get_combined_threshold}
\title{Plot major allosteric sites}
\usage{
Get_combined_threshold(
  ddG1 = ddG1,
  assay1 = assay1,
  ddG2 = ddG2,
  assay2 = assay2,
  ddG3 = ddG3,
  assay3 = assay3,
  ddG4 = ddG4,
  assay4 = assay4,
  ddG5 = ddG5,
  assay5 = assay5,
  ddG6 = ddG6,
  assay6 = assay6,
  anno = anno
)
}
\arguments{
\item{ddG1}{free energy data1}

\item{assay1}{data1's name}

\item{ddG2}{free energy data2}

\item{assay2}{data2's name}

\item{ddG3}{free energy data3}

\item{assay3}{data3's name}

\item{ddG4}{free energy data4}

\item{assay4}{data4's name}

\item{ddG5}{free energy data5}

\item{assay5}{data5's name}

\item{ddG6}{free energy data6}

\item{assay6}{data6's name}

\item{anno}{anno_final3}
}
\value{
Nothing
}
\description{
This function allows you to plot major allosteric sites.
}
