% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Get_common_interface.R
\name{Get_common_interface}
\alias{Get_common_interface}
\title{Plot common binding interface residues}
\usage{
Get_common_interface(anno_input = anno_input)
}
\arguments{
\item{anno_input}{anno_input final_anno3}
}
\value{
Nothing
}
\description{
This function allows you to plot common binding interface residues.
}
