% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Get_dc_long_kuriyan_data.R
\name{Get_dc_long_kuriyan_data}
\alias{Get_dc_long_kuriyan_data}
\title{kuriyan dcast data table}
\usage{
Get_dc_long_kuriyan_data(kuriyan_list_data = kuriyan_list_data)
}
\arguments{
\item{kuriyan_list_data}{kuriyan_list_data}
}
\value{
kuriyan_dcast_data
}
\description{
This function allows you to get kuriyan dcast data table.
}
