% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Get_weighted_mean_abs_ddG_mutcount.R
\name{Get_weighted_mean_abs_ddG_mutcount}
\alias{Get_weighted_mean_abs_ddG_mutcount}
\title{calculate weighted mean absolute free energy}
\usage{
Get_weighted_mean_abs_ddG_mutcount(ddG = ddG, assay_sele = assay_sele)
}
\arguments{
\item{ddG}{free energy data}
}
\value{
Nothing
}
\description{
This function allows you to calculate weighted mean absolute free energy change.
}
