% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Get_weighted_mean_ddG.R
\name{Get_weighted_mean_ddG}
\alias{Get_weighted_mean_ddG}
\title{calculate weighted mean free energy}
\usage{
Get_weighted_mean_ddG(ddG = ddG, assay_sele = assay_sele)
}
\arguments{
\item{ddG}{free energy data}
}
\value{
weighted mean free energy changes
}
\description{
This function allows you to calculate weighted mean free energy change.
}
