% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/krasddpcams.R
\name{krasddpcams}
\alias{krasddpcams}
\title{krasddpcams}
\usage{
krasddpcams(
  base_dir = "/Users/cweng/Desktop/Chenchun_KRAS_paper/Github_updating_20231115/"
)
}
\arguments{
\item{base_dir}{Base directory for all output files (default:NB private CRG server path; change accordingly)}
}
\value{
Nothing
}
\description{
Main analysis script.
}
