% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/krasddpcams__bind_predict_fitness.R
\name{krasddpcams__bind_predict_fitness}
\alias{krasddpcams__bind_predict_fitness}
\title{prediction binding fitness from folding and binding energy change}
\usage{
krasddpcams__bind_predict_fitness(
  mochi_parameters = mochi_parameters,
  folding_energy = folding_energy,
  binding_energy = binding_energy,
  RT = 0.001987 * (273 + 30)
)
}
\arguments{
\item{mochi_parameters}{mochi_parameters}

\item{folding_energy}{folding_energy}

\item{binding_energy}{binding_energy}

\item{RT}{0.001987*(273+30)}
}
\value{
predicted binding fitness
}
\description{
This function allows you to plot fitness distribution.
}
