% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/krasddpcams__fraction_folded.R
\name{krasddpcams__fraction_folded}
\alias{krasddpcams__fraction_folded}
\title{krasddpcams__fraction_folded}
\usage{
krasddpcams__fraction_folded(folding_energy, RT = 0.001987 * (273 + 30))
}
\arguments{
\item{folding_energy}{folding energies for 2-state model (required)}

\item{RT}{constant (default:0.001987*(273+30))}
}
\value{
Fraction folded
}
\description{
Calculate fraction folded in 2-state model.
}
