% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/krasddpcams__get_background_order_fitness.R
\name{krasddpcams__get_background_order_fitness}
\alias{krasddpcams__get_background_order_fitness}
\title{A Function to get background mutations for each blocks}
\usage{
krasddpcams__get_background_order_fitness(input = input, threshold = threshold)
}
\arguments{
\item{input}{dimsum data}

\item{threshold}{count number to identify background}
}
\value{
list by order of fitness of single mutations
}
\description{
This function allows you to get background mutations for each blocks
}
