% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/krasddpcams__get_bindingresidue_bindingpartners.R
\name{krasddpcams__get_bindingresidue_bindingpartners}
\alias{krasddpcams__get_bindingresidue_bindingpartners}
\title{Get binding residues of binding partners}
\usage{
krasddpcams__get_bindingresidue_bindingpartners(input_list)
}
\arguments{
\item{input_list}{PDB file list}
}
\value{
Nothing
}
\description{
This function allows you to get binding residues of binding partners.
}
