% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/krasddpcams__get_sotorasib_pocket.R
\name{krasddpcams__get_sotorasib_pocket}
\alias{krasddpcams__get_sotorasib_pocket}
\title{To get sotorasib pocket residues}
\usage{
krasddpcams__get_sotorasib_pocket(input_file = input_file)
}
\arguments{
\item{intput_file}{drug pdb file 6oim}
}
\value{
sortorasib distance
}
\description{
This function allows you to get sotorasib pocket residues.
}
