% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/krasddpcams__merge_ddGf_fitness.R
\name{krasddpcams__merge_ddGf_fitness}
\alias{krasddpcams__merge_ddGf_fitness}
\title{A Function to plot2D folding fitness against folding ddG}
\usage{
krasddpcams__merge_ddGf_fitness(
  prediction = prediction,
  folding_ddG = folding_ddG,
  block1_dimsum_df = block1_dimsum_df,
  block2_dimsum_df = block2_dimsum_df,
  block3_dimsum_df = block3_dimsum_df,
  wt_aa_input = wt_aa_input
)
}
\arguments{
\item{prediction}{fitness data prediction}

\item{wt_aa_input}{wt_aa_input}

\item{folding}{free energy of folding}

\item{binding}{free energy of binding}

\item{binding_assay}{binding assay's name}

\item{block_sele}{block's name}
}
\value{
Nothing
}
\description{
This function allows you to plot folding fitness against folding ddG.
}
