% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/krasddpcams__minimum_drug_distances_from_PDB.R
\name{krasddpcams__minimum_drug_distances_from_PDB}
\alias{krasddpcams__minimum_drug_distances_from_PDB}
\title{minimum interchain distance to drug Function}
\usage{
krasddpcams__minimum_drug_distances_from_PDB(
  input_file,
  chain_query = "A",
  drug = "MOV"
)
}
\arguments{
\item{input_file}{PDB file}

\item{chain_query}{KRAS chain}

\item{drug}{"MOV"}
}
\value{
result_dt
}
\description{
This function allows you to calculate minimum interchain distance to drug from PDB files.
}
