% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/krasddpcams__plot2d_ddGb_ob_pre_fitness.R
\name{krasddpcams__plot2d_ddGb_ob_pre_fitness}
\alias{krasddpcams__plot2d_ddGb_ob_pre_fitness}
\title{A Function to plot2D predicted fitness against observed fitness}
\usage{
krasddpcams__plot2d_ddGb_ob_pre_fitness(
  prediction = prediction,
  block1_dimsum_df = block1_dimsum_df,
  block2_dimsum_df = block2_dimsum_df,
  block3_dimsum_df = block3_dimsum_df,
  assay_sele = assay_sele,
  wt_aa_input = wt_aa_input
)
}
\arguments{
\item{prediction}{fitness data prediction}

\item{block1_dimsum_df}{block1_dimsum_df}

\item{block2_dimsum_df}{block2_dimsum_df}

\item{block3_dimsum_df}{block3_dimsum_df}

\item{assay_sele}{assay_sele}

\item{wt_aa_input}{wt_aa_input}
}
\value{
Nothing
}
\description{
This function allows you to plot predicted fitness against observed fitness.
}
