% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/krasddpcams__plot2d_ddGf_ob_pre_fitness_perblock.R
\name{krasddpcams__plot2d_ddGf_ob_pre_fitness_perblock}
\alias{krasddpcams__plot2d_ddGf_ob_pre_fitness_perblock}
\title{A Function to plot2D folding predicted fitness against observed fitness}
\usage{
krasddpcams__plot2d_ddGf_ob_pre_fitness_perblock(
  pre_nor = pre_nor,
  phenotypen = phenotypen
)
}
\arguments{
\item{pre_nor}{pre_nor}

\item{phenotypen}{phenotypen}
}
\value{
Nothing
}
\description{
This function allows you to plot folding predicted fitness against observed fitness.
}
