% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/krasddpcams__plot3d_ddGf_ddGb_ob_pre_fitness.R
\name{krasddpcams__plot3d_ddGf_ddGb_ob_pre_fitness}
\alias{krasddpcams__plot3d_ddGf_ddGb_ob_pre_fitness}
\title{A Function to plot3D binding fitness against folding ddG and binding ddG}
\usage{
krasddpcams__plot3d_ddGf_ddGb_ob_pre_fitness(
  binding_input = binding_input,
  binding_assay = binding_assay,
  folding_assay = folding_assay,
  mochi_parameters = mochi_parameters,
  RT = RT,
  colour_scheme
)
}
\arguments{
\item{binding_input}{binding_input}

\item{binding_assay}{binding assay's name(Binding1_RAF)}

\item{folding_assay}{free energy of binding(Abundance1)}

\item{mochi_parameters}{mochi_parameters}

\item{RT}{RT}

\item{colour_scheme}{colour scheme list}
}
\value{
Nothing
}
\description{
This function allows you to plot binding fitness against folding ddG and binding ddG.
}
