% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/krasddpcams__plot_allosteric_mut_distance_2color.R
\name{krasddpcams__plot_allosteric_mut_distance_2color}
\alias{krasddpcams__plot_allosteric_mut_distance_2color}
\title{Plot allosteric mutation vs distance}
\usage{
krasddpcams__plot_allosteric_mut_distance_2color(
  weighted_mean_ddG = weighted_mean_ddG,
  ddG = ddG,
  anno = anno,
  assay_sele = assay_sele,
  rect_input = rect_input,
  colour_scheme
)
}
\arguments{
\item{weighted_mean_ddG}{weighted mean free energy data}

\item{ddG}{free energy data}

\item{anno}{structure information annotation data}

\item{rect_input}{rect_input}

\item{colour_scheme}{colour scheme list}

\item{assay_name}{assay:"RAF","PI3"...}
}
\value{
Nothing
}
\description{
This function allows you to plot binding interface mutations heatmap.
}
