% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/krasddpcams__plot_allosteric_mutations_enrichment.R
\name{krasddpcams__plot_allosteric_mutations_enrichment}
\alias{krasddpcams__plot_allosteric_mutations_enrichment}
\title{Plot allosteric mutations enrichment}
\usage{
krasddpcams__plot_allosteric_mutations_enrichment(
  ddG = ddG,
  anno = anno,
  assay_sele = assay_sele,
  threshold = threshold,
  colour_scheme
)
}
\arguments{
\item{ddG}{free energy data}

\item{anno}{structure information annotation data}

\item{assay_sele}{assay:"RAF","PI3"...}

\item{threshold}{threshold}

\item{colour_scheme}{colour scheme list}
}
\value{
Nothing
}
\description{
This function allows you to plot allosteric mutations.
}
