% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/krasddpcams__plot_beta_strand_propagation_heatmap.R
\name{krasddpcams__plot_beta_strand_propagation_heatmap}
\alias{krasddpcams__plot_beta_strand_propagation_heatmap}
\title{beta strand heatmap}
\usage{
krasddpcams__plot_beta_strand_propagation_heatmap(
  ddG = ddG,
  assay_sele = assay_sele,
  wt_aa,
  colour_scheme
)
}
\arguments{
\item{ddG}{free energy data}

\item{assay_sele}{assay_sele}

\item{wt_aa}{wt amino acid sequence}

\item{colour_scheme}{colour scheme list}
}
\value{
Nothing
}
\description{
This function allows you to plot beta strand heatmap.
}
