% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/krasddpcams__plot_cor_gr_tecan_binding.R
\name{krasddpcams__plot_cor_gr_tecan_binding}
\alias{krasddpcams__plot_cor_gr_tecan_binding}
\title{validation tecan vs selection for bindingPCA}
\usage{
krasddpcams__plot_cor_gr_tecan_binding(
  tecandata = tecandata,
  plasmidid = plasmidid,
  single = single,
  colour_scheme
)
}
\arguments{
\item{tecandata}{tecandata output}

\item{plasmidid}{plasmid id}

\item{single}{single mutation fitness from DiMSum output}

\item{colour_scheme}{colour scheme list}
}
\value{
Nothing
}
\description{
This function allows you to plot correlation between tecan culture growthrate and selection culture fitness.
}
