% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/krasddpcams__plot_ddG_all_allosteric_heatmap.R
\name{krasddpcams__plot_ddG_all_allosteric_heatmap}
\alias{krasddpcams__plot_ddG_all_allosteric_heatmap}
\title{plot chosen position ddG heatmap}
\usage{
krasddpcams__plot_ddG_all_allosteric_heatmap(
  ddG1 = ddG1,
  assay1 = assay1,
  ddG2 = ddG2,
  assay2 = assay2,
  ddG3 = ddG3,
  assay3 = assay3,
  ddG4 = ddG4,
  assay4 = assay4,
  ddG5 = ddG5,
  assay5 = assay5,
  ddG6 = ddG6,
  assay6 = assay6,
  anno = anno,
  list1 = list1,
  list2 = list2,
  list3 = list3,
  list4 = list4,
  list5 = list5,
  list6 = list6,
  wt_aa,
  colour_scheme
)
}
\arguments{
\item{ddG1}{free energy data set assay1}

\item{assay1}{assay1}

\item{ddG2}{free energy data set assay2}

\item{assay2}{assay2}

\item{ddG3}{free energy data set assay3}

\item{assay3}{assay23}

\item{anno}{anno_final2 are chosen}

\item{list1}{list1}

\item{list2}{list2}

\item{list3}{list3}

\item{list4}{list4}

\item{list5}{list5}

\item{list6}{list6}

\item{wt_aa}{wt amino acid sequence}

\item{colour_scheme}{colour scheme list}
}
\value{
Nothing
}
\description{
This function allows you to plot chosen position ddG heatmap.
}
