% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/krasddpcams__plot_ddG_fullRAF_RBD.R
\name{krasddpcams__plot_ddG_fullRAF_RBD}
\alias{krasddpcams__plot_ddG_fullRAF_RBD}
\title{Plot binding free energy changes correlation between full RAF1 and RAF1RBD bindingPCA}
\usage{
krasddpcams__plot_ddG_fullRAF_RBD(
  ddG1 = ddG1,
  ddG2 = ddG2,
  ci95_1 = ci95_1,
  ci95_2 = ci95_2,
  colour_scheme
)
}
\arguments{
\item{ddG1}{mochi output}

\item{ddG2}{mochi output}

\item{colour_scheme}{colour scheme list}
}
\value{
Nothing
}
\description{
This function allows you to plot binding free energy changes correlation between full RAF1 and RAF1RBD bindingPCA.
}
