% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/krasddpcams__plot_ddG_fullRAF_RBD_color_fig3.R
\name{krasddpcams__plot_ddG_fullRAF_RBD_color_fig3}
\alias{krasddpcams__plot_ddG_fullRAF_RBD_color_fig3}
\title{Plot binding free energy changes correlation between full RAF1 and RAF1RBD bindingPCA}
\usage{
krasddpcams__plot_ddG_fullRAF_RBD_color_fig3(
  ddG1 = ddG1,
  ddG2 = ddG2,
  ddG_RAFRBD = ddG_RAFRBD,
  anno = anno,
  assay_sele = assay_sele,
  colour_scheme
)
}
\arguments{
\item{ddG1}{mochi output}

\item{ddG2}{mochi output}

\item{ddG_RAFRBD}{free energy data ddG_RAFRBD from six binding partner model}

\item{anno}{structure information annotation data}

\item{colour_scheme}{colour scheme list}

\item{assay_name}{assay:"RAF","PI3"...}
}
\value{
Nothing
}
\description{
This function allows you to plot binding free energy changes correlation between full RAF1 and RAF1RBD bindingPCA.
}
