% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/krasddpcams__plot_density_GAP_RAF_single_gr.R
\name{krasddpcams__plot_density_GAP_RAF_single_gr}
\alias{krasddpcams__plot_density_GAP_RAF_single_gr}
\title{Plot fitness correlation between full RAF1 and RAF1RBD bindingPCA}
\usage{
krasddpcams__plot_density_GAP_RAF_single_gr(
  GAP_fitness = GAP_fitness,
  RAF_fitness = RAF_fitness,
  driven_mt1 = driven_mt1,
  driven_mt2 = driven_mt2,
  driven_mt3 = driven_mt3,
  colour_scheme
)
}
\arguments{
\item{GAP_fitness}{dimsum output}

\item{RAF_fitness}{dimsum output}

\item{driven_mt1}{driven mutation 1}

\item{driven_mt2}{driven mutation 2}

\item{driven_mt3}{driven mutation 3}

\item{colour_scheme}{colour scheme list}
}
\value{
Nothing
}
\description{
This function allows you to plot fitness correlation between full RAF1 and RAF1RBD bindingPCA.
}
