% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/krasddpcams__plot_drivenmt_wt_gr.R
\name{krasddpcams__plot_drivenmt_wt_gr}
\alias{krasddpcams__plot_drivenmt_wt_gr}
\title{Plot growth rate correlation between wt background and mt background}
\usage{
krasddpcams__plot_drivenmt_wt_gr(
  input = input,
  driven_mt = driven_mt,
  colour_r = colour_r
)
}
\arguments{
\item{input}{dimsum output}

\item{driven_mt}{driven mutation}
}
\value{
Nothing
}
\description{
This function allows you to plot fitness correlation between wt background and mt background.
}
