% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/krasddpcams__plot_figure4A_point_all_ddG_alphabeta.R
\name{krasddpcams__plot_figure4A_point_all_ddG_alphabeta}
\alias{krasddpcams__plot_figure4A_point_all_ddG_alphabeta}
\title{all ddG point plot}
\usage{
krasddpcams__plot_figure4A_point_all_ddG_alphabeta(
  ddG1 = ddG1,
  assay1 = assay1,
  ddG2 = ddG2,
  assay2 = assay2,
  ddG3 = ddG3,
  assay3 = assay3,
  ddG4 = ddG4,
  assay4 = assay4,
  ddG5 = ddG5,
  assay5 = assay5,
  ddG6 = ddG6,
  assay6 = assay6,
  ddG7 = ddG7,
  assay7 = assay7,
  anno = anno,
  rect_input = rect_input,
  rect_alpha = rect_alpha,
  colour_scheme
)
}
\arguments{
\item{ddG1}{free energy data1}

\item{assay1}{data1's name}

\item{ddG2}{free energy data2}

\item{assay2}{data2's name}

\item{ddG3}{free energy data3}

\item{assay3}{data3's name}

\item{ddG4}{free energy data4}

\item{assay4}{data4's name}

\item{ddG5}{free energy data5}

\item{assay5}{data5's name}

\item{ddG6}{free energy data6}

\item{assay6}{data6's name}

\item{ddG7}{free energy data7}

\item{assay7}{data7's name}

\item{anno}{annotation df final_distance_dc_anno_for_anno_mt_4}

\item{rect_input}{rect for beta strands}

\item{rect_alpha}{rect for alpha 1}

\item{colour_scheme}{colour scheme list}
}
\value{
Nothing
}
\description{
This function allows you to plot all ddG heatmap.
}
