% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/krasddpcams__plot_fisher_test.R
\name{krasddpcams__plot_fisher_test}
\alias{krasddpcams__plot_fisher_test}
\title{fisher's test for decrease folding ddG in the core}
\usage{
krasddpcams__plot_fisher_test(
  target_group1 = target_group1,
  other_group2 = other_group2,
  condition1 = condition1,
  condition2 = condition2,
  g1c1 = g1c1,
  g1c2 = g1c2,
  g2c1 = g2c1,
  g2c2 = g2c2,
  x_y = x_y,
  alternative = "greater"
)
}
\arguments{
\item{target_group1}{target_group1}

\item{other_group2}{other_group2}

\item{condition1}{condition1}

\item{condition2}{condition2}

\item{g1c1}{group 1 condition 1 sample size}

\item{g1c2}{group 1 condition 2 sample size}

\item{g2c1}{group 2 condition 1 sample size}

\item{g2c2}{group 2 condition 2 sample size}

\item{x_y}{x axis name and y axis name}

\item{alternative}{"greater"...}
}
\value{
Nothing
}
\description{
This function allows you to plot fisher's test
}
