% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/krasddpcams__plot_fitness_correlation_block1.R
\name{krasddpcams__plot_fitness_correlation_block1}
\alias{krasddpcams__plot_fitness_correlation_block1}
\title{Plot fitness correlation across triplicates for block1 only}
\usage{
krasddpcams__plot_fitness_correlation_block1(fitness = fitness)
}
\arguments{
\item{fitness}{dimsum output}

\item{assay_name}{assay_name}
}
\value{
Nothing
}
\description{
This function allows you to plot fitness correlation across triplicates for block1 only。
}
