% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/krasddpcams__plot_fitness_correlation_blocks.R
\name{krasddpcams__plot_fitness_correlation_blocks}
\alias{krasddpcams__plot_fitness_correlation_blocks}
\title{A Function to plot fitness correlation between replicates}
\usage{
krasddpcams__plot_fitness_correlation_blocks(
  input = input,
  assay_name = assay_name,
  colour_scheme
)
}
\arguments{
\item{input}{dimsum data}

\item{assay_name}{binding or abundance fitness}

\item{colour_scheme}{colour scheme list}
}
\value{
Nothing
}
\description{
This function allows you to lot fitness correlation between replicates
}
