% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/krasddpcams__plot_fitness_kuriyan_data.R
\name{krasddpcams__plot_fitness_kuriyan_data}
\alias{krasddpcams__plot_fitness_kuriyan_data}
\title{A Function to plot binding fitness against kuriyan data}
\usage{
krasddpcams__plot_fitness_kuriyan_data(
  fitness = fitness,
  kuriyan_data = kuriyan_data
)
}
\arguments{
\item{fitness}{fitness data}

\item{kuriyan_data}{kuriyan_data}
}
\value{
merged data binding fitness and kuriyan score
}
\description{
This function allows you to plot binding fitness against kuriyan data
}
