% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/krasddpcams__plot_heatmap_ddG.R
\name{krasddpcams__plot_heatmap_ddG}
\alias{krasddpcams__plot_heatmap_ddG}
\title{ddG heatmap Function}
\usage{
krasddpcams__plot_heatmap_ddG(
  input = input,
  assay_name = assay_name,
  wt_aa,
  colour_scheme
)
}
\arguments{
\item{input}{mochi data}

\item{assay_name}{folding or binding}

\item{wt_aa}{wt amino acid sequence}

\item{colour_scheme}{colour scheme list}
}
\value{
Nothing
}
\description{
This function allows you to plot heatmap of free energy change.
}
