% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/krasddpcams__plot_heatmap_fitness.R
\name{krasddpcams__plot_heatmap_fitness}
\alias{krasddpcams__plot_heatmap_fitness}
\title{fitness heatmap Function}
\usage{
krasddpcams__plot_heatmap_fitness(
  input,
  assay_name,
  anno_input,
  wt_aa,
  colour_scheme
)
}
\arguments{
\item{input}{dimsum data}

\item{assay_name}{folding or binding}

\item{anno_input}{annotation}

\item{wt_aa}{wt amino acid sequence}

\item{colour_scheme}{colour scheme list}
}
\value{
Nothing
}
\description{
This function allows you to plot heatmap of fitness.
}
