% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/krasddpcams__plot_heatmap_fitness_block1.R
\name{krasddpcams__plot_heatmap_fitness_block1}
\alias{krasddpcams__plot_heatmap_fitness_block1}
\title{fitness heatmap Function for block1 only}
\usage{
krasddpcams__plot_heatmap_fitness_block1(input = input, wt_aa, colour_scheme)
}
\arguments{
\item{input}{dimsum data}

\item{wt_aa}{wt amino acid sequence}

\item{colour_scheme}{colour scheme list}

\item{assay_name}{folding or binding}
}
\value{
Nothing
}
\description{
This function allows you to plot heatmap of fitness for block1.
}
