% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/krasddpcams__plot_kuriyan_cor.R
\name{krasddpcams__plot_kuriyan_cor}
\alias{krasddpcams__plot_kuriyan_cor}
\title{A Function to plot fitness correlation with kuriyan dataset}
\usage{
krasddpcams__plot_kuriyan_cor(
  kuriyan_list = kuriyan_list,
  colour_scheme,
  all_data_pos
)
}
\arguments{
\item{kuriyan_list}{kuriyan_list"kuriyan_list.RData"}

\item{colour_scheme}{colour scheme list}

\item{all_data_pos}{all data positions}
}
\value{
Nothing
}
\description{
This function allows you to plot fitness correlation with kuriyan dataset.
}
