% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/krasddpcams__plot_major_allosteric_site.R
\name{krasddpcams__plot_major_allosteric_site}
\alias{krasddpcams__plot_major_allosteric_site}
\title{Plot major allosteric sites}
\usage{
krasddpcams__plot_major_allosteric_site(
  ddG = ddG,
  anno = anno,
  assay_name = assay_name,
  allosteric_list = allosteric_list,
  rect_input = rect_input,
  colour_scheme
)
}
\arguments{
\item{ddG}{free energy data}

\item{anno}{structure information annotation data}

\item{assay_name}{assay:"RAF","PI3"...}

\item{allosteric_list}{output data with allosteric sites and ggplot}

\item{rect_input}{rect_input}

\item{colour_scheme}{colour scheme list}
}
\value{
list of allosteric sites information
}
\description{
This function allows you to plot major allosteric sites.
}
