% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/krasddpcams__plot_pre_12_ob_3_fitness_cor_all_heldout.R
\name{krasddpcams__plot_pre_12_ob_3_fitness_cor_all_heldout}
\alias{krasddpcams__plot_pre_12_ob_3_fitness_cor_all_heldout}
\title{A Function to plot predicted fitness from replicates 1&2 against observed fitness replicate 3 for all blocks}
\usage{
krasddpcams__plot_pre_12_ob_3_fitness_cor_all_heldout(
  prediction = prediction,
  observed = observed,
  assay_sele = assay_sele,
  r_y = r_y
)
}
\arguments{
\item{prediction}{prediction}

\item{observed}{observed}

\item{assay_sele}{assay_sele}

\item{r_y}{r_y}
}
\value{
Nothing
}
\description{
This function allows you to plot predicted fitness from replicates 1&2 against observed fitness replicate 3.
}
