% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/krasddpcams__plot_RAF_invitro_cor.R
\name{krasddpcams__plot_RAF_invitro_cor}
\alias{krasddpcams__plot_RAF_invitro_cor}
\title{RAF in vitro correlation Function}
\usage{
krasddpcams__plot_RAF_invitro_cor(
  input = input,
  assay_name = assay_name,
  colour_scheme
)
}
\arguments{
\item{input}{RAF1 free energy data}

\item{assay_name}{assay}

\item{colour_scheme}{colour scheme list}
}
\value{
Nothing
}
\description{
This function allows you to plot RAF1 binding free energy Mochi inferred against in vitro data.
}
