% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/krasddpcams__plot_RAL_invitro_cor_GDI.R
\name{krasddpcams__plot_RAL_invitro_cor_GDI}
\alias{krasddpcams__plot_RAL_invitro_cor_GDI}
\title{RALGDS in vitro cor Function}
\usage{
krasddpcams__plot_RAL_invitro_cor_GDI(
  ddG = ddG,
  assay_sele = assay_sele,
  colour_scheme
)
}
\arguments{
\item{ddG}{RALGDS free energy data}

\item{assay_sele}{assay_sele}

\item{colour_scheme}{colour scheme list}
}
\value{
Nothing
}
\description{
This function allows you to plot RAF1 binding free energy Mochi inferred against in vitro data.
}
