% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/krasddpcams__pos_id.R
\name{krasddpcams__pos_id}
\alias{krasddpcams__pos_id}
\title{A Function to find mutations from aa sequence}
\usage{
krasddpcams__pos_id(input, wt_aa)
}
\arguments{
\item{input}{merged dimsum output data table}

\item{wt_aa}{wt amino acid sequence}
}
\value{
data.table with mutation positions
}
\description{
This function allows you to find mutation from dimsum output data.
}
