% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/krasddpcams__scatterplot_fitness_bs.R
\name{krasddpcams__scatterplot_fitness_bs}
\alias{krasddpcams__scatterplot_fitness_bs}
\title{Scatter plot fitness of binding vs abundance}
\usage{
krasddpcams__scatterplot_fitness_bs(
  input = input,
  assay_sele = assay_sele,
  anno = anno,
  colour_scheme
)
}
\arguments{
\item{input}{all_data_pos}

\item{assay_sele}{assay_sele}

\item{anno}{anno_final3}

\item{colour_scheme}{colour scheme list}
}
\value{
Nothing
}
\description{
This function allows you to scatter plot binding fitness vs abundance fitness.
}
