% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/krasddpcams__write_MAVEdb_scoretable.R
\name{krasddpcams__write_MAVEdb_scoretable}
\alias{krasddpcams__write_MAVEdb_scoretable}
\title{write MAVEdb from DiMSum output fitness}
\usage{
krasddpcams__write_MAVEdb_scoretable(
  input = input,
  assay_sele = assay_sele,
  output = output
)
}
\arguments{
\item{input}{DiMSum output}

\item{output}{output csv}

\item{assay_name}{assay_name}
}
\value{
Nothing
}
\description{
This function allows you to write MAVEdb data.
}
