% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Maximum_ddG_structure.R
\name{Maximum_ddG_structure}
\alias{Maximum_ddG_structure}
\title{Maximum free energy to structure}
\usage{
Maximum_ddG_structure(
  ddG = ddG,
  input_PDB = input_PDB,
  chain_KRAS = chain_KRAS,
  output_PDB_file = output_PDB_file
)
}
\arguments{
\item{ddG}{free energy data}

\item{input_PDB}{PDB file}

\item{chain_KRAS}{PDB file chain of KRAS}

\item{output_PDB_file}{export PDB file}
}
\value{
Nothing
}
\description{
This function allows you to set b factor of sturcture file(PDB) as Maximum free energy changes.
}
