% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Merge_dimsum_df.R
\name{Merge_dimsum_df}
\alias{Merge_dimsum_df}
\title{A Function to merge dimsum output}
\usage{
Merge_dimsum_df(
  merge_1 = merge_1,
  merge_2 = merge_2,
  merge_3 = merge_3,
  merge_4 = merge_4,
  merge_5 = merge_5,
  merge_6 = merge_6,
  merge_7 = merge_7
)
}
\arguments{
\item{merge_1}{1st dimsum output}

\item{merge_2}{2nd dimsum output}

\item{merge_3}{3rd dimsum output}

\item{merge_4}{4th dimsum output}

\item{merge_5}{5th dimsum output}

\item{merge_6}{6th dimsum output}

\item{merge_7}{7th dimsum output}
}
\value{
Nothing
}
\description{
This function allows you to merge dimsum output.
}
