% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Nor_overlap_single_mt_fitness.R
\name{Nor_overlap_single_mt_fitness}
\alias{Nor_overlap_single_mt_fitness}
\title{normalise overlapping variants between blocks}
\usage{
Nor_overlap_single_mt_fitness(input = input)
}
\arguments{
\item{input}{dimsum data}
}
\value{
normalized single mutation data
}
\description{
This function allows you to normalise overlapping variants between blocks.
}
