% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Normalize_growthrate_fitness.R
\name{Normalize_growthrate_fitness}
\alias{Normalize_growthrate_fitness}
\title{A Function to normalize growth rates from dimsum output}
\usage{
Normalize_growthrate_fitness(
  block1_dimsum_df = block1_dimsum_df,
  block2_dimsum_df = block2_dimsum_df,
  block3_dimsum_df = block3_dimsum_df
)
}
\arguments{
\item{block1_dimsum_df}{block1's dimsum output}

\item{block2_dimsum_df}{block2's dimsum output}

\item{block3_dimsum_df}{block3's dimsum output}
}
\value{
normalized fitness
}
\description{
This function allows you to normalize growth rates of different blocks from dimsum output according to the weighted mean stop and wt growth rates.
}
