% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Normalize_growthrate_fitness_block1.R
\name{Normalize_growthrate_fitness_block1}
\alias{Normalize_growthrate_fitness_block1}
\title{read fitness input for block1 only}
\usage{
Normalize_growthrate_fitness_block1(block1_dimsum_df = block1_dimsum_df)
}
\arguments{
\item{block1_dimsum_df}{dimsum data}
}
\value{
Nothing
}
\description{
This function allows you to input fitness for block1.
}
