% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot2d_ddGf_fitness.R
\name{Plot2d_ddGf_fitness}
\alias{Plot2d_ddGf_fitness}
\title{A Function to plot2D folding fitness against folding ddG}
\usage{
Plot2d_ddGf_fitness(
  pre_nor = pre_nor,
  mochi_parameters = mochi_parameters,
  fold_n = fold_n,
  phenotypen = phenotypen,
  RT = RT,
  bin_input = bin_input
)
}
\arguments{
\item{prediction}{fitness data prediction}

\item{folding}{free energy of folding}

\item{binding}{free energy of binding}

\item{binding_assay}{binding assay's name}

\item{block_sele}{block's name}

\item{wt_aa_input}{wt_aa_input}
}
\value{
Nothing
}
\description{
This function allows you to plot folding fitness against folding ddG.
}
