% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot2d_ddGf_ob_pre_fitness_perblock.R
\name{Plot2d_ddGf_ob_pre_fitness_perblock}
\alias{Plot2d_ddGf_ob_pre_fitness_perblock}
\title{A Function to plot2D folding predicted fitness against observed fitness}
\usage{
Plot2d_ddGf_ob_pre_fitness_perblock(pre_nor = pre_nor, phenotypen = phenotypen)
}
\arguments{
\item{pre_nor}{pre_nor}

\item{phenotypen}{phenotypen}
}
\value{
Nothing
}
\description{
This function allows you to plot folding predicted fitness against observed fitness.
}
