% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_allosteric_ddG_dis_pocket_color.R
\name{Plot_allosteric_ddG_dis_pocket_color}
\alias{Plot_allosteric_ddG_dis_pocket_color}
\title{Plot allosteric mutation vs distance colored by pockets}
\usage{
Plot_allosteric_ddG_dis_pocket_color(
  weighted_mean_ddG = weighted_mean_ddG,
  ddG = ddG,
  anno = anno,
  pocket_anno = pocket_anno,
  assay_sele = assay_sele,
  pocket = pocket,
  pocket2 = pocket2
)
}
\arguments{
\item{weighted_mean_ddG}{weighted mean free energy data}

\item{ddG}{free energy data}

\item{anno}{structure information annotation data}

\item{pocket_anno}{final_distance_dc_anno_for_anno_mt_4}

\item{assay_sele}{assay:"RAF","PI3"...}

\item{pocket}{pocket name}

\item{pocket2}{pocket2 name, Sotorasib_dis}
}
\value{
Nothing
}
\description{
This function allows you to plot binding interface mutations vs distance.
}
