% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_allosteric_mutations.R
\name{Plot_allosteric_mutations}
\alias{Plot_allosteric_mutations}
\title{Plot allosteric mutations}
\usage{
Plot_allosteric_mutations(
  ddG = ddG,
  anno = anno,
  assay_sele = assay_sele,
  rect_input = rect_input,
  rect_alpha = rect_alpha
)
}
\arguments{
\item{ddG}{free energy data}

\item{anno}{structure information annotation data}

\item{rect_input}{rect_input}

\item{rect_alpha}{rect_alpha}

\item{assay_name}{assay:"RAF","PI3"...}
}
\value{
Nothing
}
\description{
This function allows you to plot allosteric mutations.
}
