% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_allosteric_mutations_enrichment.R
\name{Plot_allosteric_mutations_enrichment}
\alias{Plot_allosteric_mutations_enrichment}
\title{Plot allosteric mutations enrichment}
\usage{
Plot_allosteric_mutations_enrichment(
  ddG = ddG,
  anno = anno,
  assay_sele = assay_sele,
  threshold = threshold
)
}
\arguments{
\item{ddG}{free energy data}

\item{anno}{structure information annotation data}

\item{assay_sele}{assay:"RAF","PI3"...}

\item{threshold}{threshold}
}
\value{
Nothing
}
\description{
This function allows you to plot allosteric mutations.
}
