% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_beta_strand_propagation_heatmap.R
\name{Plot_beta_strand_propagation_heattmap}
\alias{Plot_beta_strand_propagation_heattmap}
\title{beta strand heatmap}
\usage{
Plot_beta_strand_propagation_heattmap(ddG = ddG, assay_sele = assay_sele)
}
\arguments{
\item{ddG}{free energy data}

\item{assay_sele}{assay_sele}
}
\value{
Nothing
}
\description{
This function allows you to plot beta strand heatmap.
}
