% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_binding_interface_mutation_heatmap.R
\name{Plot_binding_interface_mutation_heatmap}
\alias{Plot_binding_interface_mutation_heatmap}
\title{Plot binding interface mutation heatmap}
\usage{
Plot_binding_interface_mutation_heatmap(ddG = ddG, assay_sele = assay_sele)
}
\arguments{
\item{ddG}{free energy data}

\item{assay_sele}{assay_sele}
}
\value{
Nothing
}
\description{
This function allows you to plot binding interface mutations heatmap.
}
