% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_cor_gr_tecan_binding.R
\name{Plot_cor_gr_tecan_binding}
\alias{Plot_cor_gr_tecan_binding}
\title{validation tecan vs selection for bindingPCA}
\usage{
Plot_cor_gr_tecan_binding(
  tecandata = tecandata,
  plasmidid = plasmidid,
  single = single
)
}
\arguments{
\item{tecandata}{tecandata output}

\item{plasmidid}{plasmid id}

\item{single}{single mutation fitness from DiMSum output}
}
\value{
Nothing
}
\description{
This function allows you to plot correlation between tecan culture growthrate and selection culture fitness.
}
