% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_ddG_all_bf_heatmap.R
\name{Plot_ddG_all_bf_heatmap}
\alias{Plot_ddG_all_bf_heatmap}
\title{plot chosen position ddG heatmap}
\usage{
Plot_ddG_all_bf_heatmap(
  ddG1 = ddG1,
  assay1 = assay1,
  ddG2 = ddG2,
  assay2 = assay2,
  ddG3 = ddG3,
  assay3 = assay3,
  ddG4 = ddG4,
  assay4 = assay4,
  ddG5 = ddG5,
  assay5 = assay5,
  ddG6 = ddG6,
  assay6 = assay6,
  anno = anno,
  bind1 = bind1,
  bind2 = bind2,
  bind3 = bind3,
  bind4 = bind4,
  bind5 = bind5,
  bind6 = bind6
)
}
\arguments{
\item{ddG1}{free energy data set assay1}

\item{assay1}{assay1}

\item{ddG2}{free energy data set assay2}

\item{assay2}{assay2}

\item{ddG3}{free energy data set assay3}

\item{assay3}{assay3}

\item{ddG4}{free energy data set assay4}

\item{assay4}{assay4}

\item{ddG5}{free energy data set assay5}

\item{assay5}{assay5}

\item{ddG6}{free energy data set assay6}

\item{assay6}{assay6}

\item{anno}{anno_final3}

\item{bind1}{bind1}

\item{bind2}{bind2}

\item{bind3}{bind3}

\item{bind4}{bind4}

\item{bind5}{bind5}

\item{bind6}{bind6}
}
\value{
Nothing
}
\description{
This function allows you to plot chosen position ddG heatmap.
}
