% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_ddG_betasheet.R
\name{Plot_ddG_betasheet}
\alias{Plot_ddG_betasheet}
\title{Plot allosteric mutation vs distance}
\usage{
Plot_ddG_betasheet(
  weighted_mean_ddG = weighted_mean_ddG,
  ddG = ddG,
  anno = anno,
  assay_sele = assay_sele,
  rect_input = rect_input
)
}
\arguments{
\item{weighted_mean_ddG}{weighted mean free energy data}

\item{ddG}{free energy data}

\item{anno}{structure information annotation data}

\item{rect_input}{rect_input}

\item{assay_name}{assay:"RAF","PI3"...}
}
\value{
Nothing
}
\description{
This function allows you to plot binding interface mutations heatmap.
}
