% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_ddG_effector_bf_heatmap.R
\name{Plot_ddG_effector_bf_heatmap}
\alias{Plot_ddG_effector_bf_heatmap}
\title{plot chosen position ddG heatmap}
\usage{
Plot_ddG_effector_bf_heatmap(
  ddG1 = ddG1,
  assay1 = assay1,
  ddG2 = ddG2,
  assay2 = assay2,
  ddG3 = ddG3,
  assay3 = assay3,
  anno = anno,
  bind1 = bind1,
  bind2 = bind2,
  bind3 = bind3
)
}
\arguments{
\item{ddG1}{free energy data set assay1}

\item{assay1}{assay1}

\item{ddG2}{free energy data set assay2}

\item{assay2}{assay2}

\item{ddG3}{free energy data set assay3}

\item{assay3}{assay23}

\item{anno}{anno_final2 are chosen}

\item{bind1}{bind1, scHAmin_ligand_RAF}

\item{bind2}{bind2, scHAmin_ligand_PI3}

\item{bind3}{bind3, scHAmin_ligand_RAL}
}
\value{
Nothing
}
\description{
This function allows you to plot chosen position ddG heatmap.
}
