% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_ddG_ss.R
\name{Plot_ddG_ss}
\alias{Plot_ddG_ss}
\title{Plot allosteric mutation in secondary structure}
\usage{
Plot_ddG_ss(
  ddG = ddG,
  anno = anno,
  assay_sele = assay_sele,
  rect_input = rect_input,
  rect_alpha = rect_alpha
)
}
\arguments{
\item{ddG}{free energy data}

\item{anno}{structure information annotation data}

\item{assay_sele}{assay:"RAF","PI3"...}

\item{rect_input}{rects}

\item{rect_alpha}{rects_alpha_all}

\item{weighted_mean_ddG}{weighted mean free energy data}
}
\value{
Nothing
}
\description{
This function allows you to plot allosteric mutation in secondary structure.
}
