% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_distribution_fitness.R
\name{Plot_distribution_fitness}
\alias{Plot_distribution_fitness}
\title{fitness distribution}
\usage{
Plot_distribution_fitness(
  input = input,
  assay_name = assay_name,
  anno_input = anno_input
)
}
\arguments{
\item{input}{dimsum data}

\item{assay_name}{folding or binding}

\item{anno_input}{anno_input}
}
\value{
Nothing
}
\description{
This function allows you to plot fitness distribution.
}
