% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_drivenmt_wt_fitness_color_fig3.R
\name{Plot_drivenmt_wt_fitness_color_fig3}
\alias{Plot_drivenmt_wt_fitness_color_fig3}
\title{Plot fitness correlation between wt background and mt background}
\usage{
Plot_drivenmt_wt_fitness_color_fig3(
  input = input,
  driven_mt = driven_mt,
  ddG_RAFRBD = ddG_RAFRBD,
  anno = anno,
  assay_sele = assay_sele
)
}
\arguments{
\item{input}{dimsum output}

\item{driven_mt}{driven mutation}

\item{ddG_RAFRBD}{free energy data ddG_RAFRBD from six binding partner model}

\item{anno}{structure information annotation data}

\item{assay_name}{assay:"RAF","PI3"...}
}
\value{
Nothing
}
\description{
This function allows you to plot fitness correlation between wt background and mt background
}
