% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_fitness_cor_RAF1_full_RBD.R
\name{Plot_fitness_cor_RAF1_full_RBD}
\alias{Plot_fitness_cor_RAF1_full_RBD}
\title{Plot fitness correlation between full RAF1 and RAF1RBD bindingPCA}
\usage{
Plot_fitness_cor_RAF1_full_RBD(fitness1 = fitness1, fitness2 = fitness2)
}
\arguments{
\item{fitness1}{dimsum output}

\item{fitness2}{dimsum output}
}
\value{
Nothing
}
\description{
This function allows you to plot fitness correlation between full RAF1 and RAF1RBD bindingPCA.
}
