% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_fitness_cor_RAF1_full_RBD_color_fig3.R
\name{Plot_fitness_cor_RAF1_full_RBD_color_fig3}
\alias{Plot_fitness_cor_RAF1_full_RBD_color_fig3}
\title{Plot fitness correlation between full RAF1 and RAF1RBD bindingPCA}
\usage{
Plot_fitness_cor_RAF1_full_RBD_color_fig3(
  fitness1 = fitness1,
  fitness2 = fitness2,
  ddG_RAFRBD = ddG_RAFRBD,
  anno = anno,
  assay_sele = assay_sele
)
}
\arguments{
\item{fitness1}{dimsum output}

\item{fitness2}{dimsum output}

\item{ddG_RAFRBD}{free energy data ddG_RAFRBD from six binding partner model}

\item{anno}{structure information annotation data}

\item{assay_name}{assay:"RAF","PI3"...}
}
\value{
Nothing
}
\description{
This function allows you to plot fitness correlation between full RAF1 and RAF1RBD bindingPCA.
}
