% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_fitness_correlation_blocks.R
\name{Plot_fitness_correlation_blocks}
\alias{Plot_fitness_correlation_blocks}
\title{A Function to plot fitness correlation between blocks}
\usage{
Plot_fitness_correlation_blocks(input = input, assay_name = assay_name)
}
\arguments{
\item{input}{dimsum data}

\item{assay_name}{binding or abundance fitness}
}
\value{
Nothing
}
\description{
This function allows you to lot fitness correlation between blocks.
}
