% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_fitness_correlation_doublevssingle.R
\name{Plot_fitness_correlation_doublevssingle}
\alias{Plot_fitness_correlation_doublevssingle}
\title{A Function to plot fitness correlation between single and double mutations}
\usage{
Plot_fitness_correlation_doublevssingle(
  input = input,
  mut = mut,
  highest_fitness = highest_fitness
)
}
\arguments{
\item{input}{dimsum data}

\item{mut}{mut}

\item{highest_fitness}{highest_fitness}
}
\value{
Nothing
}
\description{
This function allows you to lot fitness correlation between single and double mutations
}
