% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_heatmap_ddG.R
\name{Plot_heatmap_ddG}
\alias{Plot_heatmap_ddG}
\title{ddG heatmap Function}
\usage{
Plot_heatmap_ddG(input = input, assay_name = assay_name)
}
\arguments{
\item{input}{mochi data}

\item{assay_name}{folding or binding}
}
\value{
Nothing
}
\description{
This function allows you to plot heatmap of free energy change.
}
