% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_heatmap_fitness.R
\name{Plot_heatmap_fitness}
\alias{Plot_heatmap_fitness}
\title{fitness heatmap Function}
\usage{
Plot_heatmap_fitness(
  input = input,
  assay_name = assay_name,
  anno_input = anno_input
)
}
\arguments{
\item{input}{dimsum data}

\item{assay_name}{folding or binding}
}
\value{
Nothing
}
\description{
This function allows you to plot heatmap of fitness.
}
