% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_heatmap_fitness_block1.R
\name{Plot_heatmap_fitness_block1}
\alias{Plot_heatmap_fitness_block1}
\title{fitness heatmap Function for block1 only}
\usage{
Plot_heatmap_fitness_block1(input = input)
}
\arguments{
\item{input}{dimsum data}

\item{assay_name}{folding or binding}
}
\value{
Nothing
}
\description{
This function allows you to plot heatmap of fitness for block1.
}
