% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_kuriyan_cor.R
\name{Plot_kuriyan_cor}
\alias{Plot_kuriyan_cor}
\title{A Function to plot fitness correlation with kuriyan dataset}
\usage{
Plot_kuriyan_cor(kuriyan_list = kuriyan_list)
}
\arguments{
\item{kuriyan_list}{kuriyan_list"kuriyan_list.RData"}
}
\value{
Nothing
}
\description{
This function allows you to plot fitness correlation with kuriyan dataset.
}
