% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_pre_12_ob_3_fitness_cor.R
\name{Plot_pre_12_ob_3_fitness_cor}
\alias{Plot_pre_12_ob_3_fitness_cor}
\title{A Function to plot predicted fitness from replicates 1&2 against observed fitness replicate 3}
\usage{
Plot_pre_12_ob_3_fitness_cor(
  prediction = prediction,
  observed = observed,
  phenotypen = phenotypen,
  r_y = r_y
)
}
\arguments{
\item{prediction}{prediction}

\item{observed}{observed}

\item{phenotypen}{phenotypen}
}
\value{
Nothing
}
\description{
This function allows you to plot predicted fitness from replicates 1&2 against observed fitness replicate 3.
}
