% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_pre_12_ob_3_fitness_cor_all.R
\name{Plot_pre_12_ob_3_fitness_cor_all}
\alias{Plot_pre_12_ob_3_fitness_cor_all}
\title{A Function to plot predicted fitness from replicates 1&2 against observed fitness replicate 3 for all blocks}
\usage{
Plot_pre_12_ob_3_fitness_cor_all(
  prediction = prediction,
  observed = observed,
  assay_sele = assay_sele,
  r_y = r_y
)
}
\arguments{
\item{prediction}{prediction}

\item{observed}{observed}

\item{r_y}{r_y}

\item{phenotypen}{phenotypen}
}
\value{
Nothing
}
\description{
This function allows you to plot predicted fitness from replicates 1&2 against observed fitness replicate 3.
}
