% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_RAF_invitro_cor.R
\name{Plot_RAF_invitro_cor}
\alias{Plot_RAF_invitro_cor}
\title{RAF in vitro correlation Function}
\usage{
Plot_RAF_invitro_cor(input = input, assay_name = assay_name)
}
\arguments{
\item{input}{RAF1 free energy data}

\item{assay_name}{assay}
}
\value{
Nothing
}
\description{
This function allows you to plot RAF1 binding free energy Mochi inferred against in vitro data.
}
