% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_RAL_invitro_cor_ITC.R
\name{Plot_RAL_invitro_cor_ITC}
\alias{Plot_RAL_invitro_cor_ITC}
\title{RALGDS in vitro(ITC) cor Function}
\usage{
Plot_RAL_invitro_cor_ITC(ddG = ddG, assay_sele = assay_sele)
}
\arguments{
\item{ddG}{RALGDS free energy data}

\item{assay_sele}{assay_sele}
}
\value{
Nothing
}
\description{
This function allows you to plot RAF1 binding free energy Mochi inferred against in vitro data.
}
