% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_ROCAUC_ddG_fitness_binding_interface.R
\name{Plot_ROCAUC_ddG_fitness_binding_interface}
\alias{Plot_ROCAUC_ddG_fitness_binding_interface}
\title{Plot ROC-AUC for predicting contact residues by ddG or fitness}
\usage{
Plot_ROCAUC_ddG_fitness_binding_interface(
  fitness_all = fitness_all,
  weighted_meab_abs_ddGf = weighted_meab_abs_ddGf,
  weighted_meab_abs_ddGb = weighted_meab_abs_ddGb,
  anno_input = anno_input,
  bind = bind,
  bind_ligand = bind_ligand
)
}
\arguments{
\item{fitness_all}{fitness_all}

\item{weighted_meab_abs_ddGf}{weighted_meab_abs_ddGf}

\item{weighted_meab_abs_ddGb}{weighted_meab_abs_ddGb}

\item{anno_input}{anno_input}

\item{bind}{bind}

\item{bind_ligand}{bind_ligand}
}
\value{
Nothing
}
\description{
This function allows you to plot ROC-AUC for predicting contact residues by ddG or fitness
}
