% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_scatter_ddGb_ddGf.R
\name{Plot_scatter_ddGb_ddGf}
\alias{Plot_scatter_ddGb_ddGf}
\title{A Function to plot binding ddG against folding ddG colored by binding interface}
\usage{
Plot_scatter_ddGb_ddGf(
  ddG1 = ddG1,
  assay1_sele = assay1_sele,
  ddG2 = ddG2,
  assay2_sele = assay2_sele,
  anno = anno
)
}
\arguments{
\item{ddG1}{free energy data1}

\item{ddG2}{free energy data2}

\item{anno}{all_distance_anno}

\item{assay1}{data1's name}

\item{assay2}{data2's name}
}
\value{
Nothing
}
\description{
This function allows you to plot binding ddG against folding ddG
}
