% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pos_id.R
\name{Pos_id}
\alias{Pos_id}
\title{A Function to find mutations from aa sequence}
\usage{
Pos_id(input = input, wt_aa = wt_aa)
}
\arguments{
\item{input}{merged dimsum output data table}

\item{wt_aa}{wt amino acid sequence}
}
\value{
data.table with mutation positions
}
\description{
This function allows you to find mutation from dimsum output data.
}
