% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Read_ddG.R
\name{Read_ddG}
\alias{Read_ddG}
\title{import ddG file}
\usage{
Read_ddG(ddG = ddG, assay_sele = assay_sele)
}
\arguments{
\item{ddG}{free energy data}

\item{assay_sele}{assay_sele}
}
\value{
ddG data
}
\description{
This function allows you to get ddG file.
}
