% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Scatterplot_fitness_bs.R
\name{Scatterplot_fitness_bs}
\alias{Scatterplot_fitness_bs}
\title{Scatter plot fitness of binding vs abundance}
\usage{
Scatterplot_fitness_bs(input = input, assay_sele = assay_sele, anno = anno)
}
\arguments{
\item{input}{all_data_pos}

\item{assay_sele}{assay_sele}

\item{anno}{anno_final3}
}
\value{
Nothing
}
\description{
This function allows you to scatter plot binding fitness vs abundance fitness.
}
