% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Weighted_mean_ddG_structure.R
\name{Weighted_mean_ddG_structure}
\alias{Weighted_mean_ddG_structure}
\title{weighted mean free energy to structure}
\usage{
Weighted_mean_ddG_structure(
  ddG = ddG,
  input_PDB = input_PDB,
  chain_KRAS = chain_KRAS,
  Pos_correction = 0,
  output_PDB_file = output_PDB_file
)
}
\arguments{
\item{input_PDB}{PDB file}

\item{chain_KRAS}{PDB file chain of KRAS}

\item{output_PDB_file}{export PDB file}

\item{weighted_mean_ddG}{free energy data}
}
\value{
Nothing
}
\description{
This function allows you to set b factor of sturcture file(PDB) as weighted mean free energy changes.
}
