% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Weighted_mean_ddG_structure_fromweighteddata.R
\name{Weighted_mean_ddG_structure_fromweighteddata}
\alias{Weighted_mean_ddG_structure_fromweighteddata}
\title{weighted mean free energy to structure}
\usage{
Weighted_mean_ddG_structure_fromweighteddata(
  weighted_mean_ddG = weighted_mean_ddG,
  input_PDB = input_PDB,
  chain_KRAS = chain_KRAS,
  Pos_correction = 0,
  output_PDB_file = output_PDB_file
)
}
\arguments{
\item{weighted_mean_ddG}{free energy data}

\item{input_PDB}{PDB file}

\item{chain_KRAS}{PDB file chain of KRAS}

\item{output_PDB_file}{export PDB file}
}
\value{
Nothing
}
\description{
This function allows you to set b factor of sturcture file(PDB) as weighted mean free energy changes.
}
