% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Write_MAVEdb_scoretable_ddG.R
\name{Write_MAVEdb_scoretable_ddG}
\alias{Write_MAVEdb_scoretable_ddG}
\title{write MAVEdb from Mochi output ddG}
\usage{
Write_MAVEdb_scoretable_ddG(
  input = input,
  assay_name = assay_name,
  output = output
)
}
\arguments{
\item{input}{MoCHI output}

\item{assay_name}{assay_name}

\item{output}{output csv}
}
\value{
Nothing
}
\description{
This function allows you to write MAVEdb data.
}
